everion/requirements.txt
-f /usr/share/pip-wheels
appdirs==1.4.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
idna==2.6
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
